scraping/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
pydantic-settings==2.4.0

SQLAlchemy==2.0.35

python-dotenv==1.0.1

google-generativeai==0.8.3
psycopg2-binary
google-search-results
openai

